+++ Tests/KeyboardKitTests/Extensions/String+CasingTests.swift
@@ -1,5 +1,5 @@
 //
-//  String+UppercasedTests.swift
+//  String+CasingTests.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2019-07-05.
@@ -10,7 +10,7 @@ import Quick
 import Nimble
 import KeyboardKit
 
-class String_UppercasedTests: QuickSpec {
+class String_CasingTests: QuickSpec {
     
     override func spec() {
 
@@ -38,51 +38,6 @@ class String_UppercasedTests: QuickSpec {
                 expect("FOOBAR".isUppercased).to(beTrue())
                 expect("fooBar".isUppercased).to(beFalse())
                 expect("123".isUppercased).to(beFalse())
-            }
-        }
-
-        describe("uppercasing string array") {
-            
-            it("uppercases all children") {
-                let array = ["a", "b", "c"]
-                let expected = ["A", "B", "C"]
-                expect(array.uppercased()).to(equal(expected))
-            }
-        }
-        
-        describe("uppercasing string array with string arrays") {
-            
-            it("uppercases all children") {
-                let array = [["a", "b"], ["c"]]
-                let expected = [["A", "B"], ["C"]]
-                expect(array.uppercased()).to(equal(expected))
-            }
-        }
-
-        describe("case adjusted for text") {
-
-            func result(for string: String, text: String) -> String {
-                string.caseAdjusted(for: text)
-            }
-
-            it("is capitalized for single uppercased char") {
-                let result = result(for: "testing this", text: "T")
-                expect(result).to(equal("Testing This"))
-            }
-
-            it("is capitalized for capitalized text") {
-                let result = result(for: "testing this", text: "Test")
-                expect(result).to(equal("Testing This"))
-            }
-
-            it("is uppercased for uppercased text") {
-                let result = result(for: "testing this", text: "TEST")
-                expect(result).to(equal("TESTING THIS"))
-            }
-
-            it("is lowercased for lowercased text") {
-                let result = result(for: "TESTING THIS", text: "test")
-                expect(result).to(equal("testing this"))
             }
         }
     }